run: initial import of the bash script to execute BingScraper
This script may be executed by cron and will generate an email report
at the end of execution.  The user must have preconfigured smtp to send
via whichever email service they wish to use, and the user must also replace
the EMAIL_HERE text with the email address they wish to send to.