scraper: retrieve the number of account credits and print a report
This patch adds functionality at the end of the script execution such that a
report on the number of credits earned for each account.  This makes it easier
such that the user can easily see how many credits were earned in the session,
as well as their total number of credits available and lifetime credits.