scraper: use two instances to search PC/mobile and retrieve account credit info
Instead of using one instance and selectively deleting/creating another instance
to switch between PC and mobile searches, use two instances: one configured for
PC searching and accessing credit info, and another specifically configured for
mobile searches.  This minimizes the number of instances created, which leads to
less problems because selenium doesn't like creating a bunch of instances of
FireFox (at least on Windows).  This also leads to less confusion since the
credit information cannot be parsed on the mobile configuration, which is
usually the last to be loaded since mobile searches are preformed last.