scraper: logout of an account instead of closing the web browser to switch accts
In order to effectively switch accounts, logout of the Microsoft account instead
of closing the browser.  Sometimes, the browser will cache things and the only
reliable effective means of switching accounts is to explicitly logout of one
account and into another.  This also has the side effect of reducing resource
utilization in creating new browser instances.