run: modify the script to allow for running in standalone terminal mode
Previously, the script was configured explicitly for running as a cron task.
While useful, it is more useful to have the script configured as a means of
executing the script and displaying the output in the terminal window.  In order
to modify the script for running under cron, certain parts have been commented
out and noted for modification so that running as a cron task is quick and easy.