run: support execution on Mac OS X platforms and require Python 3
The BingScraper script is written for use with Python 3, and has been tested
only with Python 3.  I have no intention of supporting previous versions of
Python, so requiring version 3 shouldn't be much of a surprise.  Additionally,
I do have to do something different on Mac OS X, because the Python executable
is usually named 'python3' when installed via Homebrew.