scraper: add a user-configurable refresh interval when retrieving credit report
During the process of retrieving the credit report, I noticed the need for a
variable refresh interval especially when running on a Raspberry Pi.  The Pi
is powerful for what it is, however when running in a headless manner rendering
web pages in memory is quite resource intensive and thus the need for a variable
delay is necessary so that the page may be fully rendered before attempting to
use elements on it.