scraper: add a user-configurable browser timeout option
This option allows the user to specify the amount of time to wait for a
connection to be made to the web browser instance before giving up.  I have
noticed on the Raspberry Pi that it takes longer than usual to connect to a new
web browser instance, most likely due to the fact that the browser is being
rendered in memory which sucks virtually all of the resources at hand.

Allowing for a user-configurable timeout allows the script to be run easily on
a platform with limited resources, such as the Raspberry Pi.