scraper: use an XPath query to locate the 'Sign In' button during login
Microsoft recently changed the layout of the account login form, and in doing so
they changed the name of the control that submits the credentials to the web
server.  The new form uses a cryptic generated value for the element ID, which
is not scriptable at all.  Fortunately, XPath allows the driver to search the
page for the text of the 'Sign In' button, and until Microsoft decides to change
things again, this should be sufficient.